+++ modules/task_1/tyrina_a_deicstra_algorithm/deicstra_algorithm.h
@@ -5,11 +5,11 @@
 #include <string>
 #include <vector>
 
-std::vector<std::vector<int>> getRandomVector(const size_t count);
+std::vector<std::vector<int>> getRandomVector(const int count);
 
 std::vector<int> getDeicstra(const std::vector<std::vector<int>>& graf,
-                             const size_t count, const size_t top);
+                             const int count, const int top);
 std::vector<int> getSeqDeicstra(const std::vector<std::vector<int>>& graf,
-                                const size_t count);
+                                const int count);
 
 #endif  // MODULES_TASK_1_TYRINA_A_DEICSTRA_ALGORITHM_DEICSTRA_ALGORITHM_H_
